SquareImageRadioSelect now have their own chunk.
